+++ client/src/components/dashboard/Activity.js
@@ -1,19 +1,19 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import DHeader from './DHeader';
+import DashboardHeader from './DashboardHeader';
 
-export default function Help({ handleClick }) {
+export default function Activity({ handleClick }) {
   return (
     <>
-      <DHeader title='Help' handleClick={handleClick} />
+      <DashboardHeader title='Activity' handleClick={handleClick} />
       <div className='h-64 flex justify-center items-center text-gray-600'>
-        Help FAQ: Coming Soon...
+        Activity: Coming Soon...
       </div>
     </>
   );
 }
 
-Help.propTypes = {
+Activity.propTypes = {
   handleClick: PropTypes.func.isRequired
 };
