+++ client/src/components/dashboard/Calendar.js
@@ -1,19 +1,19 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import DHeader from './DHeader';
+import DashboardHeader from './DashboardHeader';
 
-export default function Activity({ handleClick }) {
+export default function Calendar({ handleClick }) {
   return (
     <>
-      <DHeader title='Activity' handleClick={handleClick} />
+      <DashboardHeader title='Calendar' handleClick={handleClick} />
       <div className='h-64 flex justify-center items-center text-gray-600'>
-        Activity: Coming Soon...
+        Calendar: Coming Soon...
       </div>
     </>
   );
 }
 
-Activity.propTypes = {
+Calendar.propTypes = {
   handleClick: PropTypes.func.isRequired
 };
