+++ client/src/components/dashboard/organization/Events.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import DHeader from '../index/DHeader';
+import DashboardHeader from '../DashboardHeader';
 import EventCard from '../../EventCard';
 import Badge from '../../../lib/Badge';
 
@@ -10,7 +10,7 @@ import _events from '../../../constants/events.json';
 export default function OrgEvents({ handleClick }) {
   return (
     <>
-      <DHeader
+      <DashboardHeader
         title='Events'
         subtitle='helpinghandllc'
         handleClick={handleClick}
