+++ client/src/pages/organization/Notifications.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Header from '../../components/Header';
+import Header from '../../lib/Header';
 
 export default function Notifications() {
   return (
